requirements.txt
alabaster==0.6.3
altgraph==0.10.2
amqp==1.0.13
anyjson==0.3.3
Babel==1.3
bdist-mpkg==0.5.0
bonjour-py==0.3
bottle==0.12.9
ccx2==0.2.0
celery==2.5.1
click==5.1
Django==1.8.2
django-celery==2.4.2
django-kombu==0.9.4
django-picklefield==0.3.1
docutils==0.12
Jinja2==2.7.3
kombu==2.5.16
lockfile==0.11.0
macholib==1.5.1
MarkupSafe==0.23
matplotlib==1.3.1
modulegraph==0.10.4
numpy==1.8.0rc1
pathlib==1.0.1
pbr==0.11.0
Pillow==3.0.0
platformio==2.4.0.dev0
py2app==0.7.3
Pygments==2.0.2
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
pypaths==0.1.2
pyserial==2.7
pyte==0.4.5
python-daemon==1.5.5
python-dateutil==1.5
pytz==2013.7
requests==2.8.1
scipy==0.13.0b1
scons==2.3.6
six==1.4.1
snowballstemmer==1.2.0
South==0.7.3
Sphinx==1.3b3
sphinx-rtd-theme==0.1.6
thefuck==1.7
tornado==2.3
vboxapi==1.0
wheel==0.26.0
xattr==0.6.4
zope.interface==4.1.1
